[Refactor] reducer 이름 변경한 거 반영 #387
