+++ app/src/pages/mypage/components/MyPage/Slide.tsx
@@ -7,7 +7,7 @@ import { shadow } from '../../../../styles';
 import { useParams, useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { RootState } from '../../../../redux/root';
-import { spec_create } from '../../../../redux/modal/reducer';
+import { spec_create } from '../../../../redux/specModal/reducer';
 import { ENDPOINT } from '../../../../utils/ENDPOINT';
 import setInterceptors from '../../../../apis/common/setInterceptors';
 import { customAxios } from '../../../../apis/instance';
