+++ app/src/pages/mypage/components/Spec/Visit.tsx
@@ -15,7 +15,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { AppDispatch, RootState } from '../../../../redux/root';
 import { getDetailSpec, Init_Form } from '../../../../redux/spec/reducer';
 import { Btn } from '../../../../components';
-import { spec_visit } from '../../../../redux/modal/reducer';
+import { spec_visit } from '../../../../redux/specModal/reducer';
 import { ENDPOINT } from '../../../../utils/ENDPOINT';
 import setInterceptors from '../../../../apis/common/setInterceptors';
 import { customAxios } from '../../../../apis/instance';
